requirements.txt
# Core data
pandas>=2.0.0
numpy>=1.24.0

# ENTSO-E API client
entsoe-py>=0.6.0

# HTTP requests (for regelleistung.net)
requests>=2.31.0

# Modeling
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
kaleido>=0.2.1

# Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
